TestRunner_Django/requirements.txt
# Django 及其扩展
Django==4.2.11
djangorestframework==3.15.2
django-cors-headers==4.4.0
django-filter==24.2
django-jazzmin==3.0.1
djangorestframework-simplejwt==5.3.1
djangorestframework-jwt==1.11.0
drf-yasg==1.21.7
drf-nested-routers==0.94.1
django-restql==0.15.4
django-simple-captcha==0.6.0
django-ranged-response==0.2.0
django-comment-migrate==0.1.7
django-simpleui==2023.12.12
django-import-export==3.3.4



# Celery 相关
celery==5.4.0
django-celery-beat==2.7.0
django-celery-results==2.5.1
kombu==5.4.2
amqp==5.3.1
vine==5.1.0
billiard==4.2.1

# 数据库和缓存
mysqlclient==2.1.1
redis==5.2.1
#psycopg2==2.9.9
SQLAlchemy==2.0.38
channels==4.1.0
channels-redis==4.2.0
async-timeout==4.0.3
pymysql==1.1.1

# 测试相关
pytest==7.4.4
pytest-html==3.2.0
pytest-metadata==3.1.1
httprunner==4.3.5
locust==2.25.0

# API 文档
coreapi==2.3.3
coreschema==0.0.4
uritemplate==4.1.1
inflection==0.5.1

# 工具包
python-dotenv==1.0.1
PyYAML==6.0.2
requests==2.32.3
loguru==0.4.1
pydantic==1.8.2
Pillow==10.4.0
aiohttp==3.8.5
pytz==2024.2
tzlocal==5.2
pypinyin==0.51.0
openpyxl==3.1.5
Markdown==3.7
gunicorn==22.0.0
uvicorn==0.30.3
whitenoise==6.7.0
user-agents==2.2.0
websockets==11.0.3

# 开发工具
black==22.12.0
pyinstaller==6.9.0

# 基础依赖
asgiref==3.8.1
certifi==2024.12.14
charset-normalizer==3.4.1
idna==3.10
Jinja2==3.1.5
MarkupSafe==2.1.5
sqlparse==0.5.3
typing-extensions==4.12.2
tzdata==2024.2
urllib3==1.26.18 